(Changed)[Improvement Day] Moved Delegates out of PXOneTapViewController (#2704)
* chore: Moved Delegates out of PXOneTapViewController

* chore: Remove unused comment
